Start refactoring DatabaseWork and friends.
Object browser output is looking a bit broken but is at least being
properly generated asynchronously now (instead of having the GUI
simply send the work to the db connection and wait). Still a lot of
cleaning up to do (sorting out the *Model types and separating models
and views properly, I think) but it's looking much better.