Remove DatabaseCommandWork which was being misused.
Originally this was an easy way for something to queue up several
commands to be executed. The reality is, work objects generally need
an owner to clean up after them, and to provide the notification for
that, some sort of custom implementation is needed.

This class was actually being used as a template implementation of
executing an immediate command, which wasn't really too useful. So
that is provided as a DatabaseWork method now.