Refactor ObjectBrowserWork to *not* be a DatabaseWork
Now we have ObjectBrowserDatabaseWork which is a DatabaseWork and
delegates Execute to a contained ObjectBrowserWork instance. Only the
ObjectBrowserWork is sent back to ObjectBrowser in the GUI
thread. That event handler now no longer has to call isDone() before
deleting the work, since it is disconnected from the database worker
thread.