Delete database work from the database worker thread after completion.
This ensures that database work is actually deleted afterwards,
because the work can't delete itself. Before, only the object browser
work was really deleted correctly, because there the object browser
could reap them from the GUI thread.

Along the way, remove the mutex, condition and Finished() method from
DatabaseWork itself. Also remove SynchronisableDatabaseWork which was
only used for disconnecting, and add a mutex and condition in the
database connection so that CloseSync() can wait for a worker thread
to exit.