Make the worker thread a member of database connection.
It had to be a pointer when it was a detached thread, but now it's
joinable there's no need, and this removes the need to create/delete
it.

Also, move the workQueue member into DatabaseConnection. It was a bit
ridiculous that it was in a separate place to its guarding
mutex/condition objects anyway. This also means the queue lives on if
a worker disconnects before it is all completed (e.g. if the
connection goes bad)