Simple support for showing results as grids now.
Each result goes into a separate tab in the results notebook: wxGrid
is a wxScrolledWindow and it gets confused if we try to dynamically
add them all into a single wxScrolledWindow inside the notebook page.

Formatting data as specific types is currently commented out as the
default grid table base doesn't implement setting the values, even
though the grid will display them.

Since all the results are buffered in memory and displayed in one go,
this performs horribly with large result sets. The grid auto-size
(which takes a lot of time) is only performed for the first 500 rows.