Refactor to make QueryResults a proper class.
QueryResults is constructed from a PGresult* and acts like a vector of
QueryResults::Row objects, each of which acts like a vector of
wxStrings. Refactored most of the ReadNNNN functions to parse values
to be methods of QueryResults::Row. QueryResults also remembers the
fields of the result set (ResultField refactored to
QueryResults::Field) and columns can be extracted from rows by name.

When passing a QueryResults object directly from the database thread
to the UI thread for a script result, use std::auto_ptr to ensure it
gets deallocated, as it is now allocated dynamically.

DoQuery et al from DatabaseWork now throw exceptions in case of
problems executing queries, these aren't handled yet.