Have LookupSymbolItem return a non-OK wxTreeItemId if no symbol table found.
If LookupSymbolItem is called on a database like template1, before the
system schemas have been added, then there will not actually be a
symbol table for the database yet. So in this case it should behave as
if the requested oid were not found as normal, rather than a spurious
assertion.

Triggered by selecting 'template1' and starting off object finder, and
choosing an object from the system schemas.