Improve fetching SSL information about a connection.
The SSL information is now encapsulated in an SSLInfo object, which is
created on the database worker thread. This avoids passing the SSL
object itself outside the thread, which was arguably wrong, and also
packages up the information extracted from the SSL structures.

Show the server's distinguished name before the cipher name in the
server item label. The distinguished name is kept as both the
OpenSSL-formatted string, and a vector of (name,value) pairs for
possible better formatting later.