Load schemas and extensions explicitly when loading database schema.
Previously, the existence of empty schemas was inferred by doing a
right-join of the schemas with the relations, and the schema models
were loaded independently by each database object type.

However, as SchemaModel begins to acquire attributes, having to
duplicate fetching the schema information is becoming painful. So load
the extensions and schemas explicitly, and store them in an internal
lookup table, and join them up to produce the same data model as
before. This means all the database-side joins against pg_namespace
and pg_extension can go.